Simplify tests for splits in negative test data
We propagate the split of negative samples in
[aas-core3.0-testgen 08d3ec37c] into unserializable and invalid negative
examples. This makes the generation and future maintenance of the test
code easier in this SDK as well, since we now do not have to track the
individual negative cases.

In addition, we improve the error messages if the recorded test data is
missing to make it clearer for the programmer what needs to be fixed in
that case.

[aas-core3.0-testgen 08d3ec37c]: https://github.com/aas-core-works/aas-core3.0-testgen/commit/08d3ec37c