Simplify test case for XML newline robustness (#40)
When updating to V3.0.2 of the meta-model, we noticed that the test
cases had to be changed substantially as we relied on the hard-coded AAS
models.

In this patch, we simplify the tests so that we still test for the edge
cases (robustness to newlines in XML de-serialization), but the code is
made more resilient to future meta-model updates.