Generate SDK from local meta model copy
In order to uncouple meta model, code, and test generation, we now opt
for explicitly moving the current meta model for which the SDK is
generated to the dev_scripts. This breaks the dependency and even allows
us to perform changes to the meta, if they have not been merged to
aas_core_meta.